requirements.txt
Flask==2.3.2
PyJWT==2.7.0
cryptography==41.0.3
waitress==2.1.2
Flask-HTTPAuth==4.8.0